Move our hid facade into kaleidoscope::hid
