+++ src/kaleidoscope/hid.h
@@ -1,0 +1,45 @@
+#pragma once
+#include <Arduino.h>
+
+namespace kaleidoscope::hid {
+
+// A facade on top of our HID implementation
+
+void initializeKeyboard();
+
+void pressKey(Key mappedKey);
+void releaseKey(Key mappedKey);
+void releaseAllKeys();
+void pressRawKey(Key mappedKey);
+void releaseRawKey(Key mappedKey);
+/** Flushes any pending regular key switch events and sends them out */
+void sendKeyboardReport();
+
+boolean isModifierKeyActive(Key mappedKey);
+
+void initializeConsumerControl();
+
+void pressConsumerControl(Key mappedKey);
+void releaseConsumerControl(Key mappedKey);
+
+void initializeSystemControl();
+
+void pressSystemControl(Key mappedKey);
+void releaseSystemControl(Key mappedKey);
+
+void initializeMouse();
+
+void moveMouse(signed char x, signed char y, signed char wheel);
+void clickMouseButtons(uint8_t buttons);
+void pressMouseButtons(uint8_t buttons);
+void releaseMouseButtons(uint8_t buttons);
+
+void initializeAbsoluteMouse();
+
+void moveAbsoluteMouse(signed char x, signed char y, signed char wheel);
+void moveAbsoluteMouseTo(uint16_t x, uint16_t y, signed char wheel);
+void clickAbsoluteMouseButtons(uint8_t buttons);
+void pressAbsoluteMouseButtons(uint8_t buttons);
+void releaseAbsoluteMouseButtons(uint8_t buttons);
+
+};
